refactor(memory): split canonical record helpers for governance budget
